+++ src/lib.rs
@@ -76,7 +76,7 @@ mod serialize;
 mod serializer;
 mod unpretty;
 mod valueaccess;
-mod xmlname;
+pub mod xmlname;
 mod xmlvalue;
 mod xotdata;